requirements.txt
antlr4-python3-runtime==4.7
typing==3.6.2; python_version < '3.5'
